refactor: 방 생성 로직 RoomEventService에서 RoomService로 이동
- RoomCreateEvent 처리 메서드를 RoomEventService에서 RoomService로 이동
- RoomCreateConsumer가 RoomEventService 대신 RoomService를 사용하도록 수정
- 관련 메서드 및 이벤트 발행 로직 재배치
- 불필요한 import 및 코드 정리
